requirements.txt
flask == 2.1.1
flask-sqlalchemy == 2.5.1
mysqlclient == 2.1.0
flask-login == 0.6.0
Flask-WTF == 1.0.1
Flask_Babel == 2.0.0
Flask-Mail==0.9.1
pyotp == 2.6.0
jinja2 == 3.1.1
python-slugify == 6.1.1
Flask-Minify == 0.39
Flask-QRcode == 3.0.0
email-validator == 1.2.1
Pillow == 9.1.1
pip==22.1
boto3==1.24.20
pycryptodome==3.15.0
requests==2.28.1
Flask-Migrate==3.1.0
Flask-Share==0.1.1
pyjwt==2.4.0
flask_restful==0.3.9
flask-paginate==2022.1.8
random-password-generator==2.2.0
pytz==2022.2.1
redis==4.3.4
rq==1.11.0
python3-nmap==1.5.4
pyvmomi==7.0.3
ansible==6.6.0
ansible-runner==2.3.1
